-Adjusts conversion from electric field to channel voltage to new event structure. -There is now only one module for this job. It is the equivalent to the old efieldToVoltageConverterPerChannel, with the added capability to handle CR events -The electric field class now also stores signal direction
