+++ NuRadioReco/detector/detector.py
@@ -229,7 +229,13 @@ class Detector(object):
     def get_number_of_channels(self, station_id):
         res = self.__get_channels(station_id)
         return len(res)
-
+        
+    def get_channel_ids(self, station_id):
+        channel_ids = []
+        for channel in self.__get_channels(station_id).values():
+            channel_ids.append(channel['channel_id'])
+        return channel_ids
+    
     def get_parallel_channels(self, station_id):
         res = self.__get_channels(station_id)
         orientations = np.zeros((len(res), 4))
